+++ Libraries/const.h
@@ -6,6 +6,11 @@
 #define IR_SAMPLE_PERIOD_MS     20U
 #define MAIN_LOOP_DELAY_MS      1U
 
+/* I2C timing */
+#define I2C_INIT_DELAY_MS       10U
+
+
+
 #define DMA_BUFFER_ADDRESS 0x30000000
 #define DMA_CACHE_LINE_SIZE 32  /**< STM32H7 D-Cache line size */