+++ Libraries/soccer.c
@@ -1,5 +1,6 @@
 #include "soccer.h"
 #include "mpu6050.h"
+#include "motors.h"
 
 static State_t state;
 
@@ -13,4 +14,31 @@ void updateData() {
   
   /* Update DMP attitude estimation */
   MPU6050_DMP_Update();
+}
+
+void soccer_ProcessData(const ModuleData_t* data) {  
+  // Get Euler angles from DMP
+  EulerAngles_t euler;
+  if (MPU6050_DMP_GetEulerAngles(&euler)) {
+    float yaw = euler.yaw;
+    // float roll = euler.roll;
+    // float pitch = euler.pitch;
+
+    // for (MtrID_t mtr_id = MTR0; mtr_id < 4; ++mtr_id) {
+    //   mtr_SetDecayMode(mtr_id, SLOW_DECAY);
+    // }
+    int8_t spd = 40;
+    if (yaw > 10) {
+      mtrs_Set4Speed(-spd, -spd, -spd, -spd);
+    } else if (yaw < -10) {
+      mtrs_Set4Speed(spd, spd, spd, spd);
+    } else {
+      // chase ball
+      if (IR_IsDataReady() && data->irData->ballAngle >= 0) {
+        polarMove(data->irData->ballAngle, spd + 10);
+      } else {
+        mtrs_StopAll();
+      }
+    }
+  }
 }